requirements.txt
langchain
langchain-community
langchain-google-vertexai
google-cloud-aiplatform
fastapi
uvicorn
jinja2
aiofiles
PyPDF2
pypdf
pydub
python-multipart
python-pptx
elevenlabs
